requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
backoff==2.2.1
bitarray==3.1.1
botocore==1.37.14
cached-property==1.5.2
cachetools==5.5.2
cchecksum==0.0.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==1.0.2
click==8.1.8
cryptography==44.0.2
cytoolz==1.0.1
distro==1.9.0
ecdsa==0.19.1
eth-account==0.11.3
eth-bloom==2.0.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.4.0
eth-pydantic-types==0.1.3
eth-rlp==1.0.1
eth-tester==0.9.1b1
eth-typing==3.5.2
eth-utils==2.3.2
eth_abi==5.2.0
eth_retry==0.2.1
evm-trace==0.1.3
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
futureproof==0.3.1
gmpy2==2.2.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-kms==3.4.1
googleapis-common-protos==1.69.2
gql==3.4.1
gradio==5.21.0
gradio_client==1.7.2
graphql-core==3.2.4
groovy==0.1.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.20
langchain-core==0.3.45
langchain-openai==0.3.9
langchain-text-splitters==0.3.6
langgraph==0.3.16
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.15
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
msgspec==0.19.0
multicall==0.9.3
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.66.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.8
py-ecc==6.0.0
py-evm==0.7.0a4
py-geth==5.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
pyethash==0.1.27
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rlp==3.0.0
rpds-py==0.23.1
rsa==4.9
ruff==0.11.0
safe-eth-py==6.0.0b42
safe-pysha3==1.0.4
safehttpx==0.1.6
semantic-version==2.10.0
setuptools==70.0.0
shellingham==1.5.4
sigfig==1.3.19
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.39
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tomlkit==0.13.2
toolz==1.0.0
tqdm==4.67.1
tqdm-loggable==0.2
trie==2.2.0
typer==0.15.2
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.24.0
web3==6.14.0
web3-ethereum-defi==0.28.1
web3-google-hsm==0.1.0
websockets==11.0.3
yarl==1.18.3
zstandard==0.23.0
